[iOS] [12584] working on `SearchViewController` revamp
