+++ OpenStack Summit/OpenStack Summit/TableViewHeaderView.swift
@@ -12,5 +12,5 @@ final class TableViewHeaderView: UITableViewHeaderFooterView {
     
     static let resuseIdentifier = "TableViewHeaderView"
     
-    @IBOutlet weak var titleLabel: UILabel!
+    @IBOutlet private(set) weak var titleLabel: UILabel!
 }
